issue with navbuttons, they won't switch colors properly
